refactor: hide storeAnything widget
